scripts/requirements.txt
requests>=2.28.0
PyYAML>=6.0
beautifulsoup4>=4.11.0
lxml>=4.9.0